[Jobs] utilize stable branch display to show stability. Filter out successful branches except root in display.
