+++ src/components/ProjectCell/styles.scss
@@ -1,0 +1,3 @@
+@import '../../styles/bootstrap';
+
+
