+++ src/model/index.ts
@@ -28,3 +28,11 @@ export const jobIsRoot = (job: Job): boolean => {
     'master']
   return !!names.find((name) => name == job.name)
 }
+
+/**
+ * Returns stable branches count
+ */
+export const stableBranches = (job: FolderJob): number => {
+  return job.jobs
+    .reduce((sum, current) => current.lastBuild && current.lastBuild.result == 'SUCCESS' ? sum + 1 : sum, 0)
+}
