+++ src/styles/_bootstrap.scss
@@ -1,0 +1,1 @@
+@import 'vars';
