+++ src/styles/_vars.scss
@@ -1,0 +1,6 @@
+
+$success: #4CAF50;
+$unstable: #FFEB3B;
+$failure: #f43a24;
+$aborted: #9E9E9E;
+$unknown: #69F0AE;
